+++ js/click-handler.js
@@ -2,16 +2,19 @@ var clickCount = document.getElementById("clickCount");
 
 var clicks = 0;
 var perClick = 1;
-var clickSound = "defaultClick";
+var clickSound = "defaultClickSound";
+
+var unlockedClickSounds = [
+    "defaultClickSound"
+]
 
 document.getElementById("clicker").addEventListener("click", function() {
     clicks += perClick;
     clickCount.innerText = `Clicks: ${clicks}`;
     
     totalClicks++;
-    if(clicks > highestClicks) {
-        highestClicks = clicks;
-    }
-
+    
+    checkStats();
+    checkCases();
     playSound(clickSound);
 });