+++ js/upgrades/per-click.js
@@ -5,6 +5,7 @@ var perClickCardPrice = document.getElementById("perClickCardPrice");
 var perClickSpacer = document.getElementById("perClickSpacer");
 
 var perClickLevel = 0;
+var perClickMax = false;
 
 var perClickLevelData = [
     {level: 1, price: 50, perClick: 2},
@@ -36,6 +37,8 @@ function buyPerClick() {
         if(perClickLevel >= perClickLevelData.length) {
             perClickCard.style.display = "none";
             perClickSpacer.style.display = "none";
+
+            perClickMax = true;
         }else {
             var nextPerClickLevel = perClickLevel + 1;
             var nextPerClickPrice = perClickLevelData[perClickLevel].price;
@@ -69,6 +72,8 @@ function loadPerClick() {
     if(perClickLevel >= perClickLevelData.length) {
         perClickCard.style.display = "none";
         perClickSpacer.style.display = "none";
+
+        perClickMax = true;
     }else {
         var nextPerClickLevel = perClickLevel + 1;
         var nextPerClickPrice = perClickLevelData[perClickLevel].price;