+++ js/util.js
@@ -1,3 +1,5 @@
+const delay = ms => new Promise(resolve => setTimeout(resolve, ms));
+
 function generateNumberBetween(min, max) {
     return Math.floor(Math.random() * (max - min + 1)) + min;
 }